+++ pages/cart_page.py
@@ -1,11 +1,17 @@
-from .base_page import BasePage
-from .locators import CartPageLocators
+import logging
+import time
+
+from selenium.common import NoSuchElementException, ElementClickInterceptedException
+
+from .login_page import LoginPage
+from .products_page import ProductsPage
+from .locators import CartPageLocators, CardProductPageLocator, ProductsPageLocators
 from .src import CartPageSrc
 from .locators import PageLocators
 from .locators import SideBarLocator
 
 
-class CartPage(BasePage):
+class CartPage(LoginPage, ProductsPage):
 
     # Checks that the page is Cart page
     def should_be_cart_page(self):
@@ -88,3 +94,113 @@ class CartPage(BasePage):
                     price == item_price
                 ), f"Товар ожидаемый, но цена {price} не соответствует ожидаемой {item_price}"
                 break
+
+    def put_or_del_products_in_packet(self, item_names):
+        """The method of adding/del products to the cart from the main page of the site"""
+        all_card_products = self.browser.find_elements(
+            *ProductsPageLocators.INVENT_ITEM
+        )
+        product_name_add_basket = []
+        for product in all_card_products:
+            product_name = product.find_element(*ProductsPageLocators.NAME_PRODUCT).text
+            if product_name in item_names:
+                try:
+                    self.element_is_located_in_element(
+                        product, ProductsPageLocators.BTN_ADD_OR_DEL_TO_BASKET
+                    ).click()
+                    product_name_add_basket.append(product_name)
+                except ElementClickInterceptedException as e:
+                    logging.error(
+                        f"Could not clicked elements on page: '{self.browser.current_url}'"
+                    )
+                    logging.exception(e.msg)
+                    return ElementClickInterceptedException
+        return product_name_add_basket
+
+    def view_and_remember_the_number_product_on_icon_basket(self):
+        """Method of getting the number of products in the basket after adding products to the basket"""
+        try:
+            num_product_in_basket = self.browser.find_element(
+                *ProductsPageLocators.NUMBER_PRODUCT_IN_BASKET
+            ).text
+        except NoSuchElementException as e:
+            logging.error(
+                f"Could not find elements on page: '{self.browser.current_url}'"
+            )
+            logging.exception(e.msg)
+            return 0
+        return int(num_product_in_basket)
+
+    def get_data_about_products_in_basket(self, num_product):
+        """Method of getting data from the cart page and comparing products in the cart and on the cart icon on the main
+        page"""
+        self.go_to_basket_page()
+        self.should_be_cart_page()
+        assert self.element_is_present(*CartPageLocators.BASKET_LIST)
+        product_in_basket = self.get_num_products_in_basket(
+            *CartPageLocators.LIST_PRODUCTS
+        )
+        assert num_product == len(
+            product_in_basket
+        ), "Incorrect number of items in the cart"
+
+    def get_id_product(self, element):
+        """Method of obtaining the product ID"""
+        product_id = (
+            element.find_element(*ProductsPageLocators.PRODUCT_ID)
+            .get_attribute("id")
+            .split("_title_link")[0]
+            .split("item_")[1]
+        )
+        return product_id
+
+    def add_or_del_product_from_card_product_page(self, product_id):
+        """Method of adding/removing products from/from the product card to/from the cart(s)"""
+        url = f"https://www.saucedemo.com/inventory-item.html?id={product_id}"
+        self.browser.get(url)
+        try:
+            self.browser.find_element(*CardProductPageLocator.ADD_BTN).click()
+        except NoSuchElementException as e:
+            self.browser.find_element(*CardProductPageLocator.DEL_BTN).click()
+            logging.error(
+                f"Could not find elements on page: '{self.browser.current_url}'"
+            )
+            logging.exception(e.msg)
+        self.browser.find_element(
+            *CardProductPageLocator.BACK_TO_THE_MAIN_PAGE_BTN
+        ).click()
+
+    def get_ids_and_put_or_del_products_in_packet_on_the_card_page(self, item_names):
+        """Метод добавления/удаления товаров с/из карточки товара в/из корзину(ы)"""
+        all_card_products = self.list_finded_item_by_name(item_names)[0]
+        ids = [self.get_id_product(product) for product in all_card_products]
+        for id in ids:
+            self.add_or_del_product_from_card_product_page(id, item_names)
+
+    def product_params_comparison_in_basket(self):
+        """The method of checking the coincidence of the parameters of the product in the basket"""
+        all_data = [
+            ("Sauce Labs Backpack", "29.99"),
+            ("Sauce Labs Bolt T-Shirt", "15.99"),
+            ("Sauce Labs Fleece Jacket", "49.99"),
+            ("Test.allTheThings() T-Shirt (Red)", "15.99"),
+        ]
+        products_in_basket = self.browser.find_elements(*CartPageLocators.PRODUCT_DIV)
+        for product in products_in_basket:
+            name_product_basket = product.find_element(
+                *CartPageLocators.NAME_PRODUCT
+            ).text
+            price_product_basket = self.clearing_characters(
+                "$", product.find_element(*CartPageLocators.PRICE_PRODUCT).text
+            )
+            assert (
+                name_product_basket,
+                price_product_basket,
+            ) in all_data, "The names or price of the products do not match"
+
+    def check_number_products_in_basket_is_zero(self):
+        """Method of checking that all products are removed from the basket"""
+        product_in_basket = self.get_num_products_in_basket(
+            *CartPageLocators.LIST_PRODUCTS
+        )
+        assert 0 == len(product_in_basket), "Incorrect number of items in the cart"
