+++ tests/test_TC_011_013_01.py
@@ -1,14 +1,11 @@
-import time
-
 import pytest
 from selenium.common import NoSuchElementException
 from pages.locators import ProductsPageLocators
-from pages.basket_page import BasketPage
+from pages.cart_page import CartPage
+from conf import URL
 
-link = "https://www.saucedemo.com/"
 
-
-class TestBasketPage:
+class TestCartPage:
     @pytest.mark.parametrize(
         "username, password, products",
         [
@@ -61,24 +58,26 @@ class TestBasketPage:
     def test_add_products_in_basket_from_main_page(
         self, d, username, password, products
     ):
-        """Тест проверки соответствия количества товаров на странице корзины и на иконке корзины на главной странице"""
-        page = BasketPage(d, link)
-        # Открывает страницу авторизации
+        """The test of checking the correspondence of the number of products on the cart page and on the basket icon on
+        the main page"""
+        page = CartPage(d, URL)
+        # Opens the authorization page
         page.open_page()
-        # Регистрация пользователя
+        # User Registration
         page.register_user(username, password)
-        # Добавление товаров с главной страницы
+        # Adding products from the main page
         page.put_or_del_products_in_packet(products)
-        # Удаление товаров с главной страницы
+        # Removing products from the main page
         page.put_or_del_products_in_packet(products)
-        # Получение количества товаров с иконки корзины
+        # Getting the number of products from the basket icon
         num_products_after_add_in_basket = (
             page.view_and_remember_the_number_product_on_icon_basket()
         )
-        # Получение данных со страницы корзины и сравнение количества товаров отображаемых в корзине и на иконке корзины
+        # Getting data from the cart page and comparing the number of products displayed in the cart and on\
+        # the cart icon
         page.get_data_about_products_in_basket(num_products_after_add_in_basket)
-        # Удаление всех товаров из корзины
-        page.clear_all_products_in_basket()
+        # Removing all items from the shopping cart
+        page.clear_cart()
 
     @pytest.mark.parametrize(
         "username, password, products",
@@ -132,31 +131,32 @@ class TestBasketPage:
     def test_add_products_in_basket_from_product_card_page(
         self, d, username, password, products
     ):
-        """Тест проверки возможности добавления и удаления товара из карточки товара"""
+        """Test to check the possibility of adding and removing an item from the product card"""
         list_del_items = [
             "Sauce Labs Backpack",
             "Sauce Labs Bolt T-Shirt",
             "Test.allTheThings() T-Shirt (Red)",
         ]
-        page = BasketPage(d, link)
-        # Открывает страницу авторизации
+        page = CartPage(d, URL)
+        # Opens the authorization page
         page.open_page()
-        # Регистрация пользователя
+        # User Registration
         page.register_user(username, password)
-        # Переход на карточку товара и добавление товара в корзину
+        # Go to the product card and add the product to the cart
         page.get_ids_and_put_or_del_products_in_packet_on_the_card_page(products)
-        # Переход на карточку товара и удаление товаров из списка list_del_items из корзины
+        # Go to the product card and delete products from the list_del_items list from the basket
         page.get_ids_and_put_or_del_products_in_packet_on_the_card_page(list_del_items)
-        # Получение количества товаров с иконки корзины
+        # Getting the number of products from the basket icon
         num_products_after_add_in_basket = (
             page.view_and_remember_the_number_product_on_icon_basket()
         )
-        # Получение данных со страницы корзины и сравнение количества товаров отображаемых в корзине и на иконке корзины
+        # Getting data from the cart page and comparing the number of products displayed in the cart and on \
+        # the cart icon
         page.get_data_about_products_in_basket(num_products_after_add_in_basket)
-        # Сравнение названия товара в корзине
+        # Comparison of the product name in the shopping cart
         page.product_params_comparison_in_basket()
-        # Удаление всех товаров из корзины
-        page.clear_all_products_in_basket()
+        # Removing all items from the shopping cart
+        page.clear_cart()
 
     @pytest.mark.parametrize(
         "username, password, products",
@@ -210,23 +210,24 @@ class TestBasketPage:
     def test_name_comparison_and_add_products_in_basket_from_product_card_page(
         self, d, username, password, products
     ):
-        """Тест проверки возможности добавления товара с карточки товара и сравнение добавленных товаров в корзину по названию и цене"""
-        page = BasketPage(d, link)
-        # Открывает страницу авторизации
+        """A test to check the possibility of adding a product from the product card and comparing the added products to
+        the cart by name and price"""
+        page = CartPage(d, URL)
+        # Opens the authorization page
         page.open_page()
-        # Регистрация пользователя
+        # User Registration
         page.register_user(username, password)
-        # Переход на карточку товара и добавление товара в корзину
+        # Go to the product card and add the product to the cart
         page.get_ids_and_put_or_del_products_in_packet_on_the_card_page(products)
-        # Получение количества товаров с иконки корзины
+        # Getting the number of products from the basket icon
         num_products_after_add_in_basket = (
             page.view_and_remember_the_number_product_on_icon_basket()
         )
         page.get_data_about_products_in_basket(num_products_after_add_in_basket)
-        # Сравнение названия товара в корзине
+        # Comparison of the product name in the shopping cart
         page.product_params_comparison_in_basket()
-        # Удаление всех товаров из корзины
-        page.clear_all_products_in_basket()
+        # Removing all items from the shopping cart
+        page.clear_cart()
 
     @pytest.mark.parametrize(
         "username, password, products",
@@ -278,16 +279,17 @@ class TestBasketPage:
         ],
     )
     def test_del_all_product_in_basket_page(self, d, username, password, products):
-        """Тест проверки возможности удаления всех товаров из корзины"""
-        page = BasketPage(d, link)
-        # Открывает страницу авторизации
+        """Test to check whether it is possible to delete all products from the basket"""
+        page = CartPage(d, URL)
+        # Opens the authorization page
         page.open_page()
-        # Регистрация пользователя
+        # User Registration
         page.register_user(username, password)
-        # Добавление товаров с главной страницы
+        # Adding products from the main page
         page.put_or_del_products_in_packet(products)
-        # Переход в корзину
+        # Go to the shopping cart
         page.click_button(*ProductsPageLocators.SHOP_CART_LINK)
-        time.sleep(4)
-        page.clear_all_products_in_basket()
+        # Clearing the basket of goods
+        page.clear_cart()
+        # Checking that all products have been removed from the cart
         page.check_number_products_in_basket_is_zero()
